refactoring: update, and refactor factories, and test utils
